+++ reactjs/src/web/components/molecules/challenges/ChallengeLog.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import { Link } from 'react-router-dom';
+import Title from '../../atoms/Title';
+
+const ChallengeObjective = (props: any) => {
+  const { challenge } = props;
+
+  return (
+    <React.Fragment>
+      <Title text="チャレンジログ" />
+    </React.Fragment>
+  );
+};
+
+export default ChallengeObjective;
